# docs/openapi.yaml
openapi: 3.0.3
info:
  title: AI Disaster Response Orchestrator API
  description: |
    Comprehensive REST API for disaster management, real-time alert processing, 
    AI-powered response planning, and multi-modal search capabilities using TiDB Serverless vector search.
  version: 1.0.0
  contact:
    name: AI Disaster Response Team
    email: support@disaster-orchestrator.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3001/api
    description: Development server
  - url: https://api.disaster-orchestrator.com/api
    description: Production server
security:
  - ApiKeyAuth: []
paths:
  /alerts:
    get:
      summary: Get all alerts
      description: Retrieve paginated list of disaster alerts with optional filtering
      tags:
        - Alerts
      parameters:
        - name: limit
          in: query
          description: Number of results per page
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: offset
          in: query
          description: Pagination offset
          schema:
            type: integer
            minimum: 0
            default: 0
        - name: severity
          in: query
          description: Filter by severity level
          schema:
            type: integer
            minimum: 1
            maximum: 4
        - name: type
          in: query
          description: Filter by alert type
          schema:
            type: string
            enum: [wildfire, flood, earthquake, cyclone, heatwave, landslide, other]
        - name: location
          in: query
          description: Filter by location
          schema:
            type: string
        - name: active
          in: query
          description: Filter active/inactive alerts
          schema:
            type: boolean
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AlertListResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '429':
          $ref: '#/components/responses/RateLimitExceeded'
        '500':
          $ref: '#/components/responses/InternalServerError'

    post:
      summary: Create new alert
      description: Create a new disaster alert
      tags:
        - Alerts
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAlertRequest'
      responses:
        '201':
          description: Alert created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateAlertResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '429':
          $ref: '#/components/responses/RateLimitExceeded'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /alerts/{id}:
    get:
      summary: Get alert by ID
      description: Retrieve detailed information about a specific alert
      tags:
        - Alerts
      parameters:
        - name: id
          in: path
          required: true
          description: Alert identifier
          schema:
            type: string
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AlertDetailResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /alerts/{id}/status:
    patch:
      summary: Update alert status
      description: Update the status of an existing alert
      tags:
        - Alerts
      parameters:
        - name: id
          in: path
          required: true
          description: Alert identifier
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateAlertStatusRequest'
      responses:
        '200':
          description: Status updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: Authorization
      description: API key authentication (Bearer token)
  schemas:
    Alert:
      type: object
      properties:
        id:
          type: string
          description: Unique alert identifier
        title:
          type: string
          description: Alert headline
        description:
          type: string
          description: Detailed alert description
        alertType:
          type: string
          enum: [wildfire, flood, earthquake, cyclone, heatwave, landslide, other]
          description: Type of disaster
        severity:
          type: integer
          minimum: 1
          maximum: 4
          description: Severity level (1=Low, 2=Medium, 3=High, 4=Critical)
        location:
          type: string
          description: Geographic location
        latitude:
          type: number
          format: double
          description: Latitude coordinate
        longitude:
          type: number
          format: double
          description: Longitude coordinate
        isActive:
          type: boolean
          description: Whether alert is currently active
        createdAt:
          type: string
          format: date-time
          description: Alert creation timestamp
        updatedAt:
          type: string
          format: date-time
          description: Last update timestamp
        source:
          type: string
          enum: [weather_api, social_media, satellite, manual]
          description: Source of the alert
        embedding:
          type: array
          items:
            type: number
          description: 1536-dimensional vector embedding
        rawData:
          type: object
          description: Additional metadata from source
      required:
        - id
        - title
        - description
        - alertType
        - severity
        - location
        - isActive
        - createdAt
        - source
    CreateAlertRequest:
      type: object
      properties:
        title:
          type: string
          minLength: 5
          maxLength: 255
        description:
          type: string
          minLength: 10
          maxLength: 2000
        alertType:
          type: string
          enum: [wildfire, flood, earthquake, cyclone, heatwave, landslide, other]
        severity:
          type: integer
          minimum: 1
          maximum: 4
        location:
          type: string
          minLength: 2
          maxLength: 255
        latitude:
          type: number
          format: double
          minimum: -90
          maximum: 90
        longitude:
          type: number
          format: double
          minimum: -180
          maximum: 180
        source:
          type: string
          enum: [weather_api, social_media, satellite, manual]
        rawData:
          type: object
      required:
        - title
        - description
        - alertType
        - severity
        - location
        - source
    UpdateAlertStatusRequest:
      type: object
      properties:
        status:
          type: string
          enum: [PENDING, PROCESSING, RESOLVED, FALSE_ALARM]
        metadata:
          type: object
          properties:
            resolvedBy:
              type: string
            resolution:
              type: string
      required:
        - status
...
    Pagination:
      type: object
      properties:
        total:
          type: integer
          description: Total number of results
        page:
          type: integer
          description: Current page number
        limit:
          type: integer
          description: Results per page
        totalPages:
          type: integer
          description: Total number of pages
    AlertListResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: array
          items:
            $ref: '#/components/schemas/Alert'
        pagination:
          $ref: '#/components/schemas/Pagination'
    AlertDetailResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          $ref: '#/components/schemas/Alert'
    CreateAlertResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
          properties:
            id:
              type: string
            message:
              type: string
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
          description: Error message
        code:
          type: string
          description: Error code
        details:
          type: object
          description: Additional error details
        timestamp:
          type: string
          format: date-time
  responses:
    BadRequest:
      description: Bad request - validation errors
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error: "Validation failed"
            code: "VALIDATION_ERROR"
            details:
              query: "Query must be between 10 and 500 characters"
            timestamp: "2025-09-15T12:00:00Z"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error: "Alert not found"
            code: "NOT_FOUND"
            timestamp: "2025-09-15T12:00:00Z"
    RateLimitExceeded:
      description: Rate limit exceeded
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error: "Too many requests"
            code: "RATE_LIMIT_EXCEEDED"
            timestamp: "2025-09-15T12:00:00Z"
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error: "Internal server error"
            code: "INTERNAL_ERROR"
            timestamp: "2025-09-15T12:00:00Z"
...